change: values section redesign
