+++ src/index.css
@@ -36,6 +36,10 @@ h1 {
   line-height: 1.1;
 }
 
+h2 {
+  margin-bottom: 0px;
+}
+
 button {
   border-radius: 8px;
   border: 1px solid transparent;
